+++ faker/__init__.py
@@ -2,6 +2,6 @@ from faker.factory import Factory
 from faker.generator import Generator
 from faker.proxy import Faker
 
-VERSION = "29.0.0"
+VERSION = "30.6.0"
 
 __all__ = ("Factory", "Generator", "Faker")