+++ tests/providers/test_job.py
@@ -8,6 +8,7 @@ from faker.providers.job.fr_FR import Provider as FrFrJobProvider
 from faker.providers.job.hu_HU import Provider as HuHuJobProvider
 from faker.providers.job.hy_AM import Provider as HyAmJobProvider
 from faker.providers.job.ja_JP import Provider as JaJpJobProvider
+from faker.providers.job.ka_GE import Provider as KaGeJobProvider
 from faker.providers.job.ko_KR import Provider as KoKrJobProvider
 from faker.providers.job.pt_BR import Provider as PtBrJobProvider
 from faker.providers.job.pt_PT import Provider as PtPtJobProvider
@@ -26,6 +27,14 @@ class TestJobProvider:
             assert faker.job() in JobProvider.jobs
 
 
+class TestAzAz:
+    """Test az_AZ job provider"""
+
+    def test_job(self, faker, num_samples):
+        for _ in range(num_samples):
+            assert faker.job() in AzAzJobProvider.jobs
+
+
 class TestCsCz:
     """Test cs_CZ job provider"""
 
@@ -36,60 +45,12 @@ class TestCsCz:
             assert job in CsCzJobProvider.jobs
 
 
-class TestAzAz:
-    """Test az_AZ job provider"""
-
-    def test_job(self, faker, num_samples):
-        for _ in range(num_samples):
-            assert faker.job() in AzAzJobProvider.jobs
-
-
-class TestJaJp:
-    """Test ja_JP job provider"""
-
-    def test_job(self, faker, num_samples):
-        for _ in range(num_samples):
-            assert faker.job() in JaJpJobProvider.jobs
-
-
-class TestKoKr:
-    """Test ko_KR job provider"""
-
-    def test_job(self, faker, num_samples):
-        for _ in range(num_samples):
-            assert faker.job() in KoKrJobProvider.jobs
-
-
-class TestHuHu:
-    """Test hu_HU job provider"""
-
-    def test_job(self, faker, num_samples):
-        for _ in range(num_samples):
-            assert faker.job() in HuHuJobProvider.jobs
-
-
-class TestHyAm:
-    """Test hy_AM job provider"""
-
-    def test_job(self, faker, num_samples):
-        for _ in range(num_samples):
-            assert faker.job() in HyAmJobProvider.jobs
-
-
 class TestDeDe:
     """Test de_DE job provider"""
 
     def test_job(self, faker, num_samples):
         for _ in range(num_samples):
             assert faker.job() in DeDeJobProvider.jobs
-
-
-class TestFrFr:
-    """Test fr_FR job provider"""
-
-    def test_job(self, faker, num_samples):
-        for _ in range(num_samples):
-            assert faker.job() in FrFrJobProvider.jobs
 
 
 class TestElGr:
@@ -108,12 +69,54 @@ class TestEsEs:
             assert faker.job() in EsEsJobProvider.jobs
 
 
-class TestPtPt:
-    """Test pt_PT job provider"""
+class TestFrFr:
+    """Test fr_FR job provider"""
 
     def test_job(self, faker, num_samples):
         for _ in range(num_samples):
-            assert faker.job() in PtPtJobProvider.jobs
+            assert faker.job() in FrFrJobProvider.jobs
+
+
+class TestHuHu:
+    """Test hu_HU job provider"""
+
+    def test_job(self, faker, num_samples):
+        for _ in range(num_samples):
+            assert faker.job() in HuHuJobProvider.jobs
+
+
+class TestHyAm:
+    """Test hy_AM job provider"""
+
+    def test_job(self, faker, num_samples):
+        for _ in range(num_samples):
+            assert faker.job() in HyAmJobProvider.jobs
+
+
+class TestJaJp:
+    """Test ja_JP job provider"""
+
+    def test_job(self, faker, num_samples):
+        for _ in range(num_samples):
+            assert faker.job() in JaJpJobProvider.jobs
+
+
+class TestKaGe:
+    """Test ka_GE job provider"""
+
+    def test_job(self, faker, num_samples):
+        for _ in range(num_samples):
+            job = faker.job()
+            assert isinstance(job, str)
+            assert job in KaGeJobProvider.jobs
+
+
+class TestKoKr:
+    """Test ko_KR job provider"""
+
+    def test_job(self, faker, num_samples):
+        for _ in range(num_samples):
+            assert faker.job() in KoKrJobProvider.jobs
 
 
 class TestPtBr:
@@ -122,6 +125,21 @@ class TestPtBr:
     def test_job(self, faker, num_samples):
         for _ in range(num_samples):
             assert faker.job() in PtBrJobProvider.jobs
+
+
+class TestPtPt:
+    """Test pt_PT job provider"""
+
+    def test_job(self, faker, num_samples):
+        for _ in range(num_samples):
+            assert faker.job() in PtPtJobProvider.jobs
+
+
+class TestRoRo:
+    """Test Ro_RO job provider"""
+
+    def test_job(self, faker, num_samples):
+        assert faker.job() in RoRoJobProvider.jobs
 
 
 class TestSkSk:
@@ -146,13 +164,6 @@ class TestTrTr:
 
     def test_job(self, faker, num_samples):
         assert faker.job() in TrTrJobProvider.jobs
-
-
-class TestRoRo:
-    """Test tr_TR job provider"""
-
-    def test_job(self, faker, num_samples):
-        assert faker.job() in RoRoJobProvider.jobs
 
 
 class TestViVn: